refactor(autosuggest): rename itemRenderer to renderer
