+++ packages/elements/src/autosuggest/helpers/utils.ts
@@ -1,6 +1,5 @@
-import { uuid } from '@refinitiv-ui/utils/uuid.js';
 import type { Item } from '../../item';
-import type { AutosuggestHighlightable, AutosuggestQuery, AutosuggestRenderer, Suggestion, AutosuggestItem } from './types';
+import { AutosuggestHighlightable, AutosuggestItem } from './types';
 
 /**
  * Check whether item can be highlighted
@@ -55,54 +54,4 @@ export const updateElementContent = (el: Item, query: string, label: string, val
   else {
     el.label = `${value}`;
   }
-};
-
-/**
- * Build item element from data object
- * @param suggestion Suggestion data
- * @param query A query data (usually string, but could be any entity )
- * @returns item
- */
-export const itemRenderer: AutosuggestRenderer = (suggestion: AutosuggestItem, query: AutosuggestQuery | null): HTMLElement => {
-  if (typeof suggestion !== 'object') {
-    const value = suggestion as string || '';
-
-    suggestion = {
-      label: value,
-      value
-    };
-  }
-
-  const {
-    type,
-    label,
-    title,
-    icon,
-    disabled,
-    value,
-    id
-  } = (suggestion as Suggestion);
-
-  const el = document.createElement('ef-item');
-  const isTextType = !type || type === 'text';
-  const elId = id || (isTextType ? uuid() : '');
-
-  if (elId) {
-    el.id = elId;
-  }
-
-  el.tabIndex = -1;
-  el.setAttribute('role', isTextType ? 'option' : 'presentation');
-  el.type = type || null;
-  el.disabled = !!disabled;
-  el.icon = icon || null;
-  el.value = value || label || '';
-
-  if (title) {
-    el.title = title;
-  }
-
-  updateElementContent(el, query as string || '', label || '', el.value);
-
-  return el;
 };
